backend/requirements.txt
fastapi[standard]>=0.115.6
uvicorn[standard]>=0.32.1
httpx>=0.28.1
supabase>=2.10.0
pydantic-settings>=2.6.1
privy-client>=0.5.0
